api/app/core/auth.py
import os
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBasic, HTTPBasicCredentials
from dotenv import load_dotenv
from secrets import compare_digest
from app.core.config import settings

load_dotenv()

security = HTTPBasic()

def get_current_user(credentials: HTTPBasicCredentials = Depends(security)):
    correct_username = settings.ADMIN_USERNAME
    correct_password = settings.ADMIN_PASSWORD

    is_correct_username = compare_digest(credentials.username, correct_username)
    is_correct_password = compare_digest(credentials.password, correct_password)

    if not (is_correct_username and is_correct_password):
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid credentials",
            headers={"WWW-Authenticate": "Basic"},
        )
    return credentials.username
