api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
SQLAlchemy==2.0.40
pydantic_settings==2.8.1